lag kjøretøyinputfelter til en egen komponent
brukes av brukers of søkers formue. Trenger ikke å ha 2 like.
